requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-dotenv==1.4.2
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
uWSGI>=2.0.19.1,<2.1
